Re-introducing input filters

+++ app/src/main/java/com/example/peter/thekitchenmenu/utils/unitofmeasure/UnitOfMeasureConstants.java
@@ -3,11 +3,11 @@ package com.example.peter.thekitchenmenu.utils.unitofmeasure;
 // TODO - convert to enum's.
 public class UnitOfMeasureConstants {
 
-    public static final int BASE_SI_UNIT_MASS = 1;
+    public static final double BASE_SI_UNIT_MASS = 1.;
     public static final double MAX_MASS = 10000.;
 
-    public static final int BASE_SI_UNIT_VOLUME = 1;
-    public static final int MAX_VOLUME = 10000;
+    public static final double BASE_SI_UNIT_VOLUME = 1.;
+    public static final double MAX_VOLUME = 10000.;
 
     public static final int BASE_UNIT_COUNT = 1;
     public static final int MINIMUM_COUNT = 1;
